services/assessment-generation/requirements.txt
flask>=2.0
flask-cors>=3.0
mongoengine>=0.24
python-dotenv>=0.21
requests>=2.28
streamlit>=1.20
langchain>=0.1.0
pypdf>=3.0.0
sentence-transformers>=2.2.0
unstructured>=0.10.0
python-magic-bin>=0.4.14
langchain-community>=0.0.16
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
langchain-huggingface>=0.0.6
openai
json5>=0.9
pdfkit>=1.0.0
docx
